inference/cpr_handler/app/requirements.txt
numpy
scikit-learn
google-cloud-storage
